+++ Fabric/src/main/java/de/markusbordihn/scraptechworkshop/mixin/FishingHookCustomRodsMixin.java
@@ -1,0 +1,57 @@
+/*
+ * Copyright 2025 Markus Bordihn
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and
+ * associated documentation files (the "Software"), to deal in the Software without restriction,
+ * including without limitation the rights to use, copy, modify, merge, publish, distribute,
+ * sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all copies or
+ * substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT
+ * NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+ * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM,
+ * DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+
+package de.markusbordihn.scraptechworkshop.mixin;
+
+import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
+import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
+import net.minecraft.world.entity.projectile.FishingHook;
+import net.minecraft.world.item.FishingRodItem;
+import net.minecraft.world.item.Item;
+import net.minecraft.world.item.ItemStack;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+
+@Mixin(FishingHook.class)
+public class FishingHookCustomRodsMixin {
+
+  @WrapOperation(
+      method = "shouldStopFishing",
+      at =
+          @At(
+              value = "INVOKE",
+              target = "Lnet/minecraft/world/item/ItemStack;is(Lnet/minecraft/world/item/Item;)Z",
+              ordinal = 0))
+  private boolean allowModdedRodsInMainhand(
+      ItemStack instance, Item item, Operation<Boolean> original) {
+    return instance.getItem() instanceof FishingRodItem || original.call(instance, item);
+  }
+
+  @WrapOperation(
+      method = "shouldStopFishing",
+      at =
+          @At(
+              value = "INVOKE",
+              target = "Lnet/minecraft/world/item/ItemStack;is(Lnet/minecraft/world/item/Item;)Z",
+              ordinal = 1))
+  private boolean allowModdedRodsInOffhand(
+      ItemStack instance, Item item, Operation<Boolean> original) {
+    return instance.getItem() instanceof FishingRodItem || original.call(instance, item);
+  }
+}